Tried it out on Abseil and RocksDB
Abseil looks really complex. It will be a challange to visualise it in
some reasonable way. RocksDB looks fine without the tests.

RockDB has many tests that use few libs. So, that's a use-case for the
graph plotting. In principle, the clusters should target that. But they
did not handle it too well. Maybe it's worth to have another visual tool
for such cases.

	modified:   Makefile
	modified:   README.md
	new file:   examples/targetgraph-Debug_rocksdb_notests.svg
	modified:   examples/targetgraph-_cmake.svg
	modified:   src/cmake_graph/script.py